requirements.txt
langchain==0.1.5
langchain-openai==0.0.5
openai==1.11.1
pdfminer.six==20231228
tiktoken==0.5.2
pinecone-client==2.2.4
streamlit==1.24.0